ISSUE #1659: Removed ActivitiesEnricher
