--- src/Domain/Activity/ActivitiesEnricher.php
@@ -1,129 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Activity;
-
-use App\Domain\Activity\Stream\ActivityPowerRepository;
-use App\Domain\Activity\Stream\ActivityStreamRepository;
-use App\Domain\Activity\Stream\StreamType;
-use App\Domain\Gear\CustomGear\CustomGearConfig;
-use App\Domain\Gear\GearRepository;
-use App\Domain\Gear\Maintenance\GearMaintenanceConfig;
-use App\Infrastructure\Exception\EntityNotFound;
-
-final class ActivitiesEnricher
-{
-    private Activities $enrichedActivities;
-    /** @var array<string, Activities> */
-    private array $activitiesPerActivityType;
-    /** @var array<string, Activity> */
-    private array $activitiesKeyedByActivityId;
-
-    public function __construct(
-        private readonly ActivityRepository $activityRepository,
-        private readonly ActivityPowerRepository $activityPowerRepository,
-        private readonly ActivityStreamRepository $activityStreamRepository,
-        private readonly ActivityTypeRepository $activityTypeRepository,
-        private readonly GearRepository $gearRepository,
-        private readonly GearMaintenanceConfig $gearMaintenanceConfig,
-        private readonly CustomGearConfig $customGearConfig,
-    ) {
-        $this->enrichedActivities = Activities::empty();
-        $this->activitiesPerActivityType = [];
-    }
-
-    private function enrichAll(): Activities
-    {
-        $maintenanceTags = $this->gearMaintenanceConfig->getAllMaintenanceTags();
-        $customGearTags = $this->customGearConfig->getAllGearTags();
-        $activities = $this->activityRepository->findAll();
-        $gears = $this->gearRepository->findAll();
-
-        foreach ($activities as $activity) {
-            if ($gearId = $activity->getGearId()) {
-                $activity->enrichWithGearName($gears->getByGearId($gearId)?->getName());
-            }
-            $activity->enrichWithBestPowerOutputs(
-                $this->activityPowerRepository->findBest($activity->getId())
-            );
-            $activity->enrichWithNormalizedPower(
-                $this->activityPowerRepository->findNormalizedPower($activity->getId())
-            );
-            $activity->enrichWithTags([
-                ...$maintenanceTags,
-                ...$customGearTags,
-            ]);
-
-            try {
-                $cadenceStream = $this->activityStreamRepository->findOneByActivityAndStreamType(
-                    activityId: $activity->getId(),
-                    streamType: StreamType::CADENCE
-                );
-
-                if (!empty($cadenceStream->getData())) {
-                    $activity->enrichWithMaxCadence(max($cadenceStream->getData()));
-                }
-            } catch (EntityNotFound) {
-            }
-            $this->activitiesKeyedByActivityId[(string) $activity->getId()] = $activity;
-        }
-
-        $this->enrichedActivities = $activities;
-
-        return $this->enrichedActivities;
-    }
-
-    public function getEnrichedActivitiesByActivityIds(ActivityIds $activityIds): Activities
-    {
-        $activities = Activities::empty();
-        if ($activityIds->isEmpty()) {
-            return $activities;
-        }
-
-        if ($this->enrichedActivities->isEmpty()) {
-            $this->enrichedActivities = $this->enrichAll();
-        }
-
-        foreach ($activityIds as $activityId) {
-            $activities->add($this->getEnrichedActivity($activityId));
-        }
-
-        return $activities;
-    }
-
-    public function getEnrichedActivities(): Activities
-    {
-        if ($this->enrichedActivities->isEmpty()) {
-            $this->enrichedActivities = $this->enrichAll();
-        }
-
-        return $this->enrichedActivities;
-    }
-
-    public function getEnrichedActivity(ActivityId $activityId): Activity
-    {
-        if ($this->enrichedActivities->isEmpty()) {
-            $this->enrichedActivities = $this->enrichAll();
-        }
-
-        return $this->activitiesKeyedByActivityId[(string) $activityId] ?? throw new EntityNotFound('Activity not found: '.$activityId);
-    }
-
-    /**
-     * @return array<string, Activities>
-     */
-    public function getActivitiesPerActivityType(): array
-    {
-        if (empty($this->activitiesPerActivityType)) {
-            $activityTypes = $this->activityTypeRepository->findAll();
-
-            /** @var ActivityType $activityType */
-            foreach ($activityTypes as $activityType) {
-                $this->activitiesPerActivityType[$activityType->value] = $this->getEnrichedActivities()->filterOnActivityType($activityType);
-            }
-        }
-
-        return $this->activitiesPerActivityType;
-    }
-}
