+++ src/Domain/Activity/ActivitySummaryRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Activity;
+
+interface ActivitySummaryRepository
+{
+    public function find(ActivityId $activityId): ActivitySummary;
+}
