+++ src/Domain/Activity/Eddington/EddingtonCalculator.php
@@ -4,16 +4,14 @@ declare(strict_types=1);
 
 namespace App\Domain\Activity\Eddington;
 
-use App\Domain\Activity\ActivitiesEnricher;
-use App\Domain\Activity\ActivityIdRepository;
+use App\Domain\Activity\ActivityRepository;
 use App\Domain\Activity\Eddington\Config\EddingtonConfiguration;
 use App\Infrastructure\ValueObject\Measurement\UnitSystem;
 
 final readonly class EddingtonCalculator
 {
     public function __construct(
-        private ActivityIdRepository $activityIdRepository,
-        private ActivitiesEnricher $activitiesEnricher,
+        private ActivityRepository $activityRepository,
         private EddingtonConfiguration $eddingtonConfiguration,
         private UnitSystem $unitSystem,
     ) {
@@ -27,13 +25,13 @@ final readonly class EddingtonCalculator
         $eddingtons = [];
         /** @var Config\EddingtonConfigItem $eddingtonConfigItem */
         foreach ($this->eddingtonConfiguration as $eddingtonConfigItem) {
-            $activityIds = $this->activityIdRepository->findBySportTypes($eddingtonConfigItem->getSportTypesToInclude());
-            if ($activityIds->isEmpty()) {
+            $activities = $this->activityRepository->findBySportTypes($eddingtonConfigItem->getSportTypesToInclude());
+            if ($activities->isEmpty()) {
                 continue;
             }
 
             $eddington = Eddington::getInstance(
-                activities: $this->activitiesEnricher->getEnrichedActivitiesByActivityIds($activityIds),
+                activities: $activities,
                 config: $eddingtonConfigItem,
                 unitSystem: $this->unitSystem
             );
