+++ src/Domain/Dashboard/Widget/ChallengeConsistencyWidget.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\Dashboard\Widget;
 
-use App\Domain\Activity\ActivitiesEnricher;
+use App\Domain\Activity\ActivityRepository;
 use App\Domain\Calendar\Months;
 use App\Domain\Challenge\Consistency\ConsistencyChallengeCalculator;
 use App\Domain\Challenge\Consistency\ConsistencyChallenges;
@@ -14,7 +14,7 @@ use Twig\Environment;
 final readonly class ChallengeConsistencyWidget implements Widget
 {
     public function __construct(
-        private ActivitiesEnricher $activitiesEnricher,
+        private ActivityRepository $activityRepository,
         private ConsistencyChallengeCalculator $consistencyChallengeCalculator,
         private Environment $twig,
     ) {
@@ -35,7 +35,7 @@ final readonly class ChallengeConsistencyWidget implements Widget
 
     public function render(SerializableDateTime $now, WidgetConfiguration $configuration): string
     {
-        $allActivities = $this->activitiesEnricher->getEnrichedActivities();
+        $allActivities = $this->activityRepository->findAll();
 
         $allMonths = Months::create(
             startDate: $allActivities->getFirstActivityStartDate(),
