+++ templates/html/dashboard/widget/widget--peak-power-outputs.html.twig
@@ -57,7 +57,7 @@
                                         {{ powerOutput.getRelativePower() }}<span class="text-xs">w/kg</span>
                                     </td>
                                     <td class="px-6 py-2 text-left truncate">
-                                        {{ powerOutput.getActivity()|activityLink() }}
+                                        {{ getActivity(powerOutput.getActivityId())|activityLink() }}
                                     </td>
                                 </tr>
                             {% endif %}
