+++ tests/Domain/Activity/DbalActivityIdRepositoryTest.php
@@ -6,13 +6,10 @@ use App\Domain\Activity\ActivityId;
 use App\Domain\Activity\ActivityIdRepository;
 use App\Domain\Activity\ActivityIds;
 use App\Domain\Activity\ActivityRepository;
-use App\Domain\Activity\ActivityType;
 use App\Domain\Activity\ActivityWithRawData;
 use App\Domain\Activity\ActivityWithRawDataRepository;
 use App\Domain\Activity\DbalActivityIdRepository;
 use App\Domain\Activity\DbalActivityWithRawDataRepository;
-use App\Domain\Activity\SportType\SportType;
-use App\Domain\Activity\SportType\SportTypes;
 use App\Domain\Gear\GearId;
 use App\Infrastructure\Exception\EntityNotFound;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
@@ -58,83 +55,6 @@ class DbalActivityIdRepositoryTest extends ContainerTestCase
         $this->assertEquals(
             ActivityIds::fromArray([$activityOne->getId(), $activityTwo->getId(), $activityThree->getId()]),
             $this->activityIdRepository->findAll()
-        );
-    }
-
-    public function testFindByStartDate(): void
-    {
-        $activityOne = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(1))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
-            ->withSportType(SportType::BADMINTON)
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityOne,
-            ['raw' => 'data']
-        ));
-        $activityTwo = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(2))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
-            ->withSportType(SportType::RUN)
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityTwo,
-            ['raw' => 'data']
-        ));
-        $activityThree = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(3))
-            ->withSportType(SportType::MOUNTAIN_BIKE_RIDE)
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityThree,
-            ['raw' => 'data']
-        ));
-
-        $this->assertEquals(
-            ActivityIds::fromArray([$activityOne->getId(), $activityTwo->getId()]),
-            $this->activityIdRepository->findByStartDate(SerializableDateTime::fromString('2023-10-10'), null)
-        );
-
-        $this->assertEquals(
-            ActivityIds::fromArray([$activityOne->getId()]),
-            $this->activityIdRepository->findByStartDate(SerializableDateTime::fromString('2023-10-10'), ActivityType::RACQUET_PADDLE_SPORTS)
-        );
-    }
-
-    public function testFindBySportTypes(): void
-    {
-        $activityOne = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(1))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
-            ->withSportType(SportType::BADMINTON)
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityOne,
-            ['raw' => 'data']
-        ));
-        $activityTwo = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(2))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
-            ->withSportType(SportType::RUN)
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityTwo,
-            ['raw' => 'data']
-        ));
-        $activityThree = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(3))
-            ->withSportType(SportType::MOUNTAIN_BIKE_RIDE)
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityThree,
-            ['raw' => 'data']
-        ));
-
-        $this->assertEquals(
-            ActivityIds::fromArray([$activityTwo->getId(), $activityThree->getId()]),
-            $this->activityIdRepository->findBySportTypes(SportTypes::fromArray([SportType::RUN, SportType::MOUNTAIN_BIKE_RIDE]))
         );
     }
