--- tests/Domain/Activity/DbalActivityRepositoryTest.php
@@ -1,118 +1,0 @@
-<?php
-
-namespace App\Tests\Domain\Activity;
-
-use App\Domain\Activity\Activity;
-use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityRepository;
-use App\Domain\Activity\ActivitySummary;
-use App\Domain\Activity\ActivityWithRawData;
-use App\Domain\Activity\ActivityWithRawDataRepository;
-use App\Domain\Activity\DbalActivityRepository;
-use App\Domain\Activity\DbalActivityWithRawDataRepository;
-use App\Infrastructure\Exception\EntityNotFound;
-use App\Infrastructure\ValueObject\Time\SerializableDateTime;
-use App\Tests\ContainerTestCase;
-use Spatie\Snapshots\MatchesSnapshots;
-
-class DbalActivityRepositoryTest extends ContainerTestCase
-{
-    use MatchesSnapshots;
-
-    private ActivityRepository $activityRepository;
-    private ActivityWithRawDataRepository $activityWithRawDataRepository;
-
-    public function testItShouldSaveAndFind(): void
-    {
-        $activity = ActivityBuilder::fromDefaults()->build();
-
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activity,
-            ['raw' => 'data']
-        ));
-
-        $persisted = $this->activityRepository->find($activity->getId());
-        $this->assertEquals(
-            $activity,
-            $persisted,
-        );
-    }
-
-    public function testFindItShouldThrowWhenNotFound(): void
-    {
-        $this->expectException(EntityNotFound::class);
-        $this->activityRepository->find(ActivityId::fromUnprefixed(1));
-    }
-
-    public function testItShouldSaveAndFindSummary(): void
-    {
-        $activity = ActivityBuilder::fromDefaults()->build();
-
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activity,
-            ['raw' => 'data']
-        ));
-
-        $persisted = $this->activityRepository->findSummary($activity->getId());
-        $this->assertEquals(
-            ActivitySummary::create(
-                name: $activity->getName(),
-                startDateTime: $activity->getStartDate(),
-            ),
-            $persisted,
-        );
-    }
-
-    public function testFindSummaryItShouldThrowWhenNotFound(): void
-    {
-        $this->expectException(EntityNotFound::class);
-        $this->activityRepository->findSummary(ActivityId::fromUnprefixed(1));
-    }
-
-    public function testFindAll(): void
-    {
-        $activityOne = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(1))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityOne,
-            ['raw' => 'data']
-        ));
-        $activityTwo = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(2))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityTwo,
-            ['raw' => 'data']
-        ));
-        $activityThree = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(3))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityThree,
-            ['raw' => 'data']
-        ));
-
-        $this->assertEquals(
-            [$activityOne->getId(), $activityTwo->getId(), $activityThree->getId()],
-            $this->activityRepository->findAll()->map(fn (Activity $activity): ActivityId => $activity->getId())
-        );
-    }
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->activityRepository = new DbalActivityRepository(
-            $this->getConnection(),
-        );
-        $this->activityWithRawDataRepository = new DbalActivityWithRawDataRepository(
-            $this->getConnection(),
-            $this->activityRepository
-        );
-    }
-}
