+++ tests/Domain/Activity/DbalEnrichedActivityRepositoryTest.php
@@ -1,0 +1,170 @@
+<?php
+
+namespace App\Tests\Domain\Activity;
+
+use App\Domain\Activity\Activity;
+use App\Domain\Activity\ActivityId;
+use App\Domain\Activity\ActivityIds;
+use App\Domain\Activity\ActivityRepository;
+use App\Domain\Activity\ActivityType;
+use App\Domain\Activity\ActivityWithRawData;
+use App\Domain\Activity\ActivityWithRawDataRepository;
+use App\Domain\Activity\DbalActivityWithRawDataRepository;
+use App\Domain\Activity\SportType\SportType;
+use App\Domain\Activity\SportType\SportTypes;
+use App\Infrastructure\Exception\EntityNotFound;
+use App\Infrastructure\ValueObject\Time\SerializableDateTime;
+use App\Tests\ContainerTestCase;
+use Spatie\Snapshots\MatchesSnapshots;
+
+class DbalEnrichedActivityRepositoryTest extends ContainerTestCase
+{
+    use MatchesSnapshots;
+
+    private ActivityRepository $activityRepository;
+    private ActivityWithRawDataRepository $activityWithRawDataRepository;
+
+    public function testItShouldSaveAndFind(): void
+    {
+        $activity = ActivityBuilder::fromDefaults()->build();
+
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activity,
+            ['raw' => 'data']
+        ));
+
+        $persisted = $this->activityRepository->find($activity->getId());
+        $this->assertEquals(
+            $activity,
+            $persisted,
+        );
+    }
+
+    public function testFindItShouldThrowWhenNotFound(): void
+    {
+        $this->expectException(EntityNotFound::class);
+        $this->activityRepository->find(ActivityId::fromUnprefixed(1));
+    }
+
+    public function testFindAll(): void
+    {
+        $activityOne = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(1))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityOne,
+            ['raw' => 'data']
+        ));
+        $activityTwo = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(2))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityTwo,
+            ['raw' => 'data']
+        ));
+        $activityThree = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(3))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityThree,
+            ['raw' => 'data']
+        ));
+
+        $this->assertEquals(
+            [$activityOne->getId(), $activityTwo->getId(), $activityThree->getId()],
+            $this->activityRepository->findAll()->map(fn (Activity $activity): ActivityId => $activity->getId())
+        );
+    }
+
+    public function testFindByStartDate(): void
+    {
+        $activityOne = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(1))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
+            ->withSportType(SportType::BADMINTON)
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityOne,
+            ['raw' => 'data']
+        ));
+        $activityTwo = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(2))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
+            ->withSportType(SportType::RUN)
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityTwo,
+            ['raw' => 'data']
+        ));
+        $activityThree = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(3))
+            ->withSportType(SportType::MOUNTAIN_BIKE_RIDE)
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityThree,
+            ['raw' => 'data']
+        ));
+
+        $this->assertEquals(
+            ActivityIds::fromArray([$activityOne->getId(), $activityTwo->getId()]),
+            $this->activityRepository->findByStartDate(SerializableDateTime::fromString('2023-10-10'), null)->map(fn (Activity $activity): ActivityId => $activity->getId())
+        );
+
+        $this->assertEquals(
+            ActivityIds::fromArray([$activityOne->getId()]),
+            $this->activityRepository->findByStartDate(SerializableDateTime::fromString('2023-10-10'), ActivityType::RACQUET_PADDLE_SPORTS)->map(fn (Activity $activity): ActivityId => $activity->getId())
+        );
+    }
+
+    public function testFindBySportTypes(): void
+    {
+        $activityOne = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(1))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
+            ->withSportType(SportType::BADMINTON)
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityOne,
+            ['raw' => 'data']
+        ));
+        $activityTwo = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(2))
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
+            ->withSportType(SportType::RUN)
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityTwo,
+            ['raw' => 'data']
+        ));
+        $activityThree = ActivityBuilder::fromDefaults()
+            ->withActivityId(ActivityId::fromUnprefixed(3))
+            ->withSportType(SportType::MOUNTAIN_BIKE_RIDE)
+            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
+            ->build();
+        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
+            $activityThree,
+            ['raw' => 'data']
+        ));
+
+        $this->assertEquals(
+            ActivityIds::fromArray([$activityTwo->getId(), $activityThree->getId()]),
+            $this->activityRepository->findBySportTypes(SportTypes::fromArray([SportType::RUN, SportType::MOUNTAIN_BIKE_RIDE]))->map(fn (Activity $activity): ActivityId => $activity->getId())
+        );
+    }
+
+    #[\Override]
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->activityRepository = $this->getContainer()->get(ActivityRepository::class);
+        $this->activityWithRawDataRepository = new DbalActivityWithRawDataRepository(
+            $this->getConnection(),
+            $this->activityRepository
+        );
+    }
+}
